Confirmed backwards compatibility with the changes made to nih.py. Data can also now we limited to minimum value.
